refactor: split up filtering chips file into multiple files
