+++ lib/launch_filtering/widgets/widgets.dart
@@ -1,2 +1,7 @@
 export 'filtering_chips.dart';
-export 'search_bar.dart';
+export 'flight_number_chip.dart';
+export 'rocket_chip.dart';
+export 'sorting_chips.dart';
+export 'successfulness_chip.dart';
+export 'time_chip.dart';
+export 'year_range_chip.dart';
